server/requirements.txt
flask
python-dotenv
requests
PyGithub
langchain
langchain-google-genai
google-generativeai
pre-commit